+++ MK4duo/Configuration_Version.h
@@ -24,7 +24,7 @@
   #define CONFIGURATION_VERSION_H
 
   #define FIRMWARE_NAME "MK4duo"
-  #define SHORT_BUILD_VERSION "4.3.11_dev"
+  #define SHORT_BUILD_VERSION "4.3.12_dev"
   #define BUILD_VERSION FIRMWARE_NAME "_" SHORT_BUILD_VERSION
   #define STRING_DISTRIBUTION_DATE __DATE__ " " __TIME__    // build date and time
   // It might also be appropriate to define a location where additional information can be found