+++ MK4duo/src/MK_Main.cpp
@@ -65,6 +65,10 @@ float progress = 0.0;
 
 uint8_t mk_debug_flags = DEBUG_NONE;
 
+#if ENABLED(LASERBEAM) || ENABLED(CNCROUTER)
+  static PrinterMode printer_mode = PRINTER_MODE_FFF;
+#endif
+
 /**
  * Cartesian Current Position
  *   Used to track the logical position as moves are queued.
@@ -1840,7 +1844,7 @@ static void clean_up_after_endstop_or_probe_move() {
 
     // When deploying make sure BLTOUCH is not already triggered
     #if ENABLED(BLTOUCH)
-      if (deploy && TEST_BLTOUCH()) { stop(); return true; }
+      //if (deploy && TEST_BLTOUCH()) { stop(); return true; }
     #elif ENABLED(Z_PROBE_SLED)
       if (axis_unhomed_error(true, false, false)) { stop(); return true; }
     #elif ENABLED(Z_PROBE_ALLEN_KEY)
@@ -2865,6 +2869,13 @@ static void homeaxis(AxisEnum axis) {
   }
 #endif
 
+#if ENABLED(CNCROUTER) && ENABLED(FAST_PWM_CNCROUTER)
+  void print_cncspeed() {
+    SERIAL_MV(" CNC speed: ", getCNCSpeed());
+    SERIAL_M(" rpm ");
+  }
+#endif
+
 #ifndef MIN_COOLING_SLOPE_DEG
   #define MIN_COOLING_SLOPE_DEG 1.50
 #endif
@@ -3381,11 +3392,11 @@ inline void gcode_G0_G1(
 
     #if ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_G1)
       if (lfire) {
-        if (code_seen('S') && IsRunning()) laser.intensity = code_value_float();
-        if (code_seen('L') && IsRunning()) laser.duration = code_value_ulong();
-        if (code_seen('P') && IsRunning()) laser.ppm = code_value_float();
-        if (code_seen('D') && IsRunning()) laser.diagnostics = code_value_bool();
-        if (code_seen('B') && IsRunning()) laser_set_mode(code_value_int());
+        if (code_seen('S')) laser.intensity = code_value_float();
+        if (code_seen('L')) laser.duration = code_value_ulong();
+        if (code_seen('P')) laser.ppm = code_value_float();
+        if (code_seen('D')) laser.diagnostics = code_value_bool();
+        if (code_seen('B')) laser_set_mode(code_value_int());
 
         laser.status = LASER_ON;
         laser.fired = LASER_FIRE_G1;
@@ -3429,7 +3440,9 @@ inline void gcode_G0_G1(
  *    G3 X20 Y12 R14   ; CCW circle with r=14 ending at X20 Y12
  */
 #if ENABLED(ARC_SUPPORT)
+
   inline void gcode_G2_G3(bool clockwise) {
+
     if (IsRunning()) {
 
       #if ENABLED(SF_ARC_FIX)
@@ -3440,11 +3453,11 @@ inline void gcode_G0_G1(
       gcode_get_destination();
 
       #if ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_G1)
-        if (code_seen('S') && IsRunning()) laser.intensity = code_value_float();
-        if (code_seen('L') && IsRunning()) laser.duration = code_value_ulong();
-        if (code_seen('P') && IsRunning()) laser.ppm = code_value_float();
-        if (code_seen('D') && IsRunning()) laser.diagnostics = code_value_bool();
-        if (code_seen('B') && IsRunning()) laser_set_mode(code_value_int());
+        if (code_seen('S')) laser.intensity = code_value_float();
+        if (code_seen('L')) laser.duration = code_value_ulong();
+        if (code_seen('P')) laser.ppm = code_value_float();
+        if (code_seen('D')) laser.diagnostics = code_value_bool();
+        if (code_seen('B')) laser_set_mode(code_value_int());
 
         laser.status = LASER_ON;
         laser.fired = LASER_FIRE_G1;
@@ -3487,6 +3500,7 @@ inline void gcode_G0_G1(
       }
     }
   }
+
 #endif // ARC_SUPPORT
 
 /**
@@ -5703,43 +5717,75 @@ inline void gcode_G92() {
   }
 #endif // ULTIPANEL
 
-#if (ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE))
+#if (ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE)) || ENABLED(CNCROUTER)
+
   /**
-   * M3: S - Setting laser beam or fire laser 
+   * M3: Setting laser beam or fire laser - CNC clockwise speed
+   * M4: Turn on laser beam or CNC counter clockwise speed
+   *      S - Laser intensity or CNC speed
+   *      L - Laser Duration
+   *      P - PPM
+   *      D - Diagnostic
+   *      B - Set mode
    */
-  inline void gcode_M3_M4() {
-    if (code_seen('S') && IsRunning()) laser.intensity = code_value_float();
-    if (code_seen('L') && IsRunning()) laser.duration = code_value_ulong();
-    if (code_seen('P') && IsRunning()) laser.ppm = code_value_float();
-    if (code_seen('D') && IsRunning()) laser.diagnostics = code_value_bool();
-    if (code_seen('B') && IsRunning()) laser_set_mode(code_value_int());
+  inline void gcode_M3_M4(bool clockwise) {
 
-    laser.status = LASER_ON;
-    laser.fired = LASER_FIRE_SPINDLE;
+    #if ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE)
+      if (printer_mode == PRINTER_MODE_LASER) {
+        if(IsRunning()) {
+          if (code_seen('S')) laser.intensity = code_value_float();
+          if (code_seen('L')) laser.duration = code_value_ulong();
+          if (code_seen('P')) laser.ppm = code_value_float();
+          if (code_seen('D')) laser.diagnostics = code_value_bool();
+          if (code_seen('B')) laser_set_mode(code_value_int());
+        }
 
-    lcd_update();
+        laser.status = LASER_ON;
+        laser.fired = LASER_FIRE_SPINDLE;
+
+        lcd_update();
+      }
+    #endif
+
+    #if ENABLED(CNCROUTER)
+      if (printer_mode == PRINTER_MODE_CNC)
+        if (code_seen('S')) setCNCRouterSpeed(code_value_ulong(), clockwise);
+    #endif
 
     prepare_move_to_destination();
   }
 
   /**
-   * M5: Turn off laser beam
+   * M5: Turn off laser beam - CNC off
    */
   inline void gcode_M5() {
-    if (laser.status != LASER_OFF) {
-      laser.status = LASER_OFF;
-      laser.mode = CONTINUOUS;
-      laser.duration = 0;
 
-      lcd_update();
+    #if ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE)
+      if (printer_mode == PRINTER_MODE_LASER) {
+        if (laser.status != LASER_OFF) {
+          laser.status = LASER_OFF;
+          laser.mode = CONTINUOUS;
+          laser.duration = 0;
 
-      prepare_move_to_destination();
+          lcd_update();
 
-      if (laser.diagnostics)
-        SERIAL_EM("Laser M5 called and laser OFF");
-    }
+          prepare_move_to_destination();
+
+          if (laser.diagnostics)
+            SERIAL_EM("Laser M5 called and laser OFF");
+        }
+      }
+    #endif
+
+    #if ENABLED(CNCROUTER)
+      if (printer_mode == PRINTER_MODE_CNC) {
+        disable_cncrouter();
+        prepare_move_to_destination();
+      }
+    #endif
   }
-#endif // LASERBEAM
+
+#endif // LASERBEAM || CNCROUTER
 
 /**
  * M17: Enable power on all stepper motors
@@ -6378,6 +6424,10 @@ inline void gcode_M81() {
     #endif
   #endif
 
+  #if ENABLED(CNCROUTER)
+    disable_cncrouter();
+  #endif
+
   safe_delay(1000); // Wait 1 second before switching off
 
   #if HAS(SUICIDE)
@@ -6725,7 +6775,7 @@ inline void gcode_M105() {
 
   GET_TARGET_HOTEND(105);
 
-  #if HAS(TEMP_0) || HAS(TEMP_BED) || ENABLED(HEATER_0_USES_MAX6675) || HAS(TEMP_COOLER) || ENABLED(FLOWMETER_SENSOR)
+  #if HAS(TEMP_0) || HAS(TEMP_BED) || ENABLED(HEATER_0_USES_MAX6675) || HAS(TEMP_COOLER) || ENABLED(FLOWMETER_SENSOR) || (ENABLED(CNCROUTER) && ENABLED(FAST_PWM_CNCROUTER))
     SERIAL_S(OK);
     #if HAS(TEMP_0) || HAS(TEMP_BED) || ENABLED(HEATER_0_USES_MAX6675)
       print_heaterstates();
@@ -6738,6 +6788,9 @@ inline void gcode_M105() {
     #endif
     #if ENABLED(FLOWMETER_SENSOR)
       print_flowratestate();
+    #endif
+    #if ENABLED(CNCROUTER) && ENABLED(FAST_PWM_CNCROUTER)
+      print_cncspeed();
     #endif
   #else // HASNT(TEMP_0) && HASNT(TEMP_BED)
     SERIAL_LM(ER, MSG_ERR_NO_THERMISTORS);
@@ -8496,6 +8549,37 @@ inline void gcode_M428() {
   }
 }
 
+#if (ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE)) || ENABLED(CNCROUTER)
+
+  /**
+   * M450: Report printer mode
+   */
+  inline void gcode_M450() {
+    SERIAL_M("PrinterMode:");
+    switch(printer_mode) {
+      case PRINTER_MODE_LASER:
+        SERIAL_M("Laser");
+      break;
+      case PRINTER_MODE_CNC:
+        SERIAL_M("CNC");
+      break;
+      default:
+        SERIAL_M("FFF");
+    }
+    SERIAL_E;
+  }
+
+  inline void gcode_M451_M452_M453(const PrinterMode mode) {
+    if (IsRunning()) SERIAL_EM("Cannot change printer mode while running");
+    else {
+      stop();
+      printer_mode = mode;
+      gcode_M450();
+    }
+  }
+
+#endif // LASERBEAM || CNCROUTER
+
 /**
  * M500: Store settings in EEPROM
  */
@@ -8914,16 +8998,19 @@ inline void gcode_M532() {
       laser.rasterlaserpower =  laser.intensity;
     }
 
-    if (code_seen('L') && IsRunning()) laser.duration = code_value_ulong();
-    if (code_seen('P') && IsRunning()) laser.ppm = code_value_float();
-    if (code_seen('B') && IsRunning()) laser_set_mode(code_value_int());
-    if (code_seen('R') && IsRunning()) laser.raster_mm_per_pulse = (code_value_float());
+    if(IsRunning()) {
+      if (code_seen('L')) laser.duration = code_value_ulong();
+      if (code_seen('P')) laser.ppm = code_value_float();
+      if (code_seen('B')) laser_set_mode(code_value_int());
+      if (code_seen('R')) laser.raster_mm_per_pulse = (code_value_float());
+    }
 
     if (code_seen('F')) {
       float next_feedrate = code_value_linear_units();
       if(next_feedrate > 0.0) feedrate_mm_s = next_feedrate;
     }
   }
+
 #endif // LASERBEAM
 
 #if MECH(MUVE3D)
@@ -10057,6 +10144,11 @@ void process_next_command() {
           gcode_G10_G11(codenum == 10); break;
       #endif // FWRETRACT
 
+      // G17 - G19: XXX CNC plane selection
+      // G17 -> XY (default)
+      // G18 -> ZX 
+      // G19 -> YZ
+
       #if ENABLED(NOZZLE_CLEAN_FEATURE)
         case 12: // G12: Nozzle Clean
           gcode_G12(); break;
@@ -10108,16 +10200,27 @@ void process_next_command() {
           break;
       #endif
 
+      // G40 Compensation Off XXX CNC
+      // G54-G59 Coordinate system selection (CNC XXX)      
+
       case 60: // G60 Saved Coordinates
         gcode_G60(); break;
       case 61: // G61 Restore Coordinates
         gcode_G61(); break;
+   
+      // G80: Cancel Canned Cycle (XXX CNC)
+
       case 90: // G90
         relative_mode = false; break;
       case 91: // G91
         relative_mode = true; break;
       case 92: // G92
         gcode_G92(); break;
+
+      // G92.x Reset Coordinate System Offset (CNC XXX)
+      // G93: Feed Rate Mode (Inverse Time Mode) (CNC XXX)
+      // G94: Feed Rate Mode (Units per Minute) (CNC XXX)
+
     }
     break;
 
@@ -10128,12 +10231,19 @@ void process_next_command() {
           gcode_M0_M1(); break;
       #endif // ULTIPANEL
 
-      #if ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE)
-        case 3: // M03: Setting laser beam
-        case 4: // M04: Turn on laser beam
-          gcode_M3_M4(); break;
-        case 5: // M05: Turn off laser beam
+      #if ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE) || ENABLED(CNCROUTER)
+        case 3: // M03: Setting laser beam or CNC clockwise speed
+        case 4: // M04: Turn on laser beam or CNC counter clockwise speed
+          gcode_M3_M4(codenum == 3); break;
+        case 5: // M05: Turn off laser beam or CNC stop
           gcode_M5(); break;
+        // case 6: // M06 - Tool change CNC XXX
+        // case 7: // M07 - Mist coolant CNC XXX
+        // case 8: // M08 - Flood coolant CNC XXX
+        // case 9: // M09 - Coolant off CNC XXX
+        // case 10: // M10 - Vacuum on CNC XXX
+        // case 11: // M11 - Vacuum off CNC XXX
+
       #endif // LASERBEAM
 
       case 17: // M17: Enable/Power all stepper motors
@@ -10529,6 +10639,21 @@ void process_next_command() {
 
       case 428: // M428 Apply current_position to home_offset
         gcode_M428(); break;
+
+      #if (ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE)) || ENABLED(CNCROUTER)
+        case 450:
+          gcode_M450(); break; // report printer mode
+        case 451:
+          gcode_M451_M452_M453(PRINTER_MODE_FFF); break;    // set printer mode printer
+        #if ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_SPINDLE)
+          case 452:
+            gcode_M451_M452_M453(PRINTER_MODE_LASER); break;  // set printer mode laser
+        #endif
+        #if ENABLED(CNCROUTER)
+          case 453:
+            gcode_M451_M452_M453(PRINTER_MODE_CNC); break;    // set printer mode router
+        #endif
+      #endif
 
       case 500: // M500: Store settings in EEPROM
         gcode_M500(); break;
@@ -11957,7 +12082,7 @@ static void report_current_position() {
   }
 #endif
 
-#if ENABLED(FAST_PWM_FAN) || ENABLED(FAST_PWM_COOLER)
+#if ENABLED(FAST_PWM_FAN) || ENABLED(FAST_PWM_COOLER) || ENABLED(FAST_PWM_CNCROUTER)
 
   void setPwmFrequency(uint8_t pin, uint8_t val) {
     val &= 0x07;
@@ -12023,6 +12148,7 @@ static void report_current_position() {
       #endif
     }
   }
+
 #endif // FAST_PWM_FAN
 
 void quickstop_stepper() {
@@ -12117,6 +12243,7 @@ void manage_inactivity(bool ignore_stepper_queue/*=false*/) {
         laser_peripherals_off();
       #endif
     #endif
+
   }
 
   #if HAS(CHDK) // Check if pin should be set to LOW after M240 set it to HIGH
@@ -12359,6 +12486,10 @@ void idle(
     flowrate_manage();
   #endif
 
+  #if ENABLED(CNCROUTER)
+    cnc_manage();
+  #endif
+
   manage_inactivity(
     #if ENABLED(FILAMENT_CHANGE_FEATURE)
       no_stepper_sleep
@@ -12404,6 +12535,10 @@ void kill(const char* lcd_msg) {
     #endif
   #endif
 
+  #if ENABLED(CNCROUTER)
+     disable_cncrouter();
+  #endif
+
   #if HAS(POWER_SWITCH)
     SET_INPUT(PS_ON_PIN);
   #endif
@@ -12439,6 +12574,10 @@ void stop() {
     #endif
   #endif
 
+  #if ENABLED(CNCROUTER)
+     disable_cncrouter();
+  #endif
+
   if (IsRunning()) {
     Running = false;
     Stopped_gcode_LastN = gcode_LastN; // Save last g_code for restart
@@ -12452,13 +12591,13 @@ void stop() {
 /**
  * Marlin entry-point: Set up before the program loop
  *  - Set up Alligator Board
- *  - Set up WANHAO D6 OLED
  *  - Set up the kill pin, filament runout, power hold
  *  - Start the serial port
  *  - Print startup messages and diagnostics
  *  - Get EEPROM or default settings
  *  - Initialize managers for:
  *    • temperature
+ *    • CNCROUTER
  *    • planner
  *    • watchdog
  *    • stepper
@@ -12539,6 +12678,10 @@ void setup() {
   SYNC_PLAN_POSITION_KINEMATIC();
 
   thermalManager.init();    // Initialize temperature loop
+
+  #if ENABLED(CNCROUTER)
+    cnc_init();
+  #endif
 
   #if ENABLED(USE_WATCHDOG)
     watchdog_init();