+++ MK4duo/src/enum.h
@@ -153,6 +153,14 @@ enum TempState {
   #endif
 #endif
 
+#if ENABLED(LASERBEAM) || ENABLED(CNCROUTER)
+  enum PrinterMode {
+    PRINTER_MODE_FFF,
+    PRINTER_MODE_LASER,
+    PRINTER_MODE_CNC
+  };
+#endif
+
 #if ENABLED(MESH_BED_LEVELING) && NOMECH(DELTA)
   enum MeshLevelingState {
     MeshReport,