+++ Documentation/changelog.md
@@ -1,3 +1,11 @@
+### Version 4.3.12 dev
+* Add CNC Router
+* Add M450 - Report Printer Mode
+* Add M451 - Select FFF Printer Mode
+* Add M452 - Select Laser Printer Mode
+* Add M453 - Select CNC Printer Mode
+* Fix and clear code.
+
 ### Version 4.3.11 dev
 * Add G38.2 and G38.3 Probe target - similar to G28 except it uses the Z_MIN endstop for all three axes
 * Fix and clear code.